# Conflicts:
#	src/main/java/dev/nbcsparta/assignment/commerce_backoffice/controller/ManagerAuthController.java

+++ src/main/java/dev/nbcsparta/assignment/commerce_backoffice/dto/CreateProductResponse.java
@@ -9,7 +9,7 @@ public record CreateProductResponse(
         Long id,
         String name,
         String category,
-        Long price,
+        int price,
         int quantity,
         ProductStatus status,
         LocalDateTime createdAt,
