+++ src/main/java/dev/nbcsparta/assignment/commerce_backoffice/dto/UpdateProductResponse.java
@@ -5,7 +5,7 @@ import dev.nbcsparta.assignment.commerce_backoffice.entity.Product;
 public record UpdateProductResponse(
         String name,
         String category,
-        Long price
+        int price
 ) {
     public static UpdateProductResponse from(Product product) {
         return new UpdateProductResponse(
